requirements.txt
oracledb
python-dotenv
Flask